Moved RuADReCT files to toloka
